requirements.txt
python-telegram-bot==13.15
python-dotenv
requests